seperation of concern
